drivers/shared: move eventer to subpackage under drivers shared package
