+++ drivers/rawexec/driver.go
@@ -12,6 +12,7 @@ import (
 	plugin "github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/client/driver/executor"
 	dstructs "github.com/hashicorp/nomad/client/driver/structs"
+	"github.com/hashicorp/nomad/drivers/shared/eventer"
 	"github.com/hashicorp/nomad/plugins/base"
 	"github.com/hashicorp/nomad/plugins/drivers"
 	"github.com/hashicorp/nomad/plugins/drivers/utils"
@@ -70,7 +71,7 @@ var (
 type RawExecDriver struct {
 	// eventer is used to handle multiplexing of TaskEvents calls such that an
 	// event can be broadcast to all callers
-	eventer *utils.Eventer
+	eventer *eventer.Eventer
 
 	// config is the driver configuration set by the SetConfig RPC
 	config *Config
@@ -122,7 +123,7 @@ func NewRawExecDriver(logger hclog.Logger) drivers.DriverPlugin {
 	ctx, cancel := context.WithCancel(context.Background())
 	logger = logger.Named(pluginName)
 	return &RawExecDriver{
-		eventer:        utils.NewEventer(ctx, logger),
+		eventer:        eventer.NewEventer(ctx, logger),
 		config:         &Config{},
 		tasks:          newTaskStore(),
 		ctx:            ctx,
