+++ drivers/shared/eventer/eventer_test.go
@@ -1,4 +1,4 @@
-package utils
+package eventer
 
 import (
 	"context"
